一つのにまとめた
